kafka, shared: Add support for pulling existing state
This commit is a bit large and makes a couple of changes, but the big
gist here is that it modifies the KafkaHandler to keep an internal
record of aggregated ObjectUpdates. This materialized view is used by
SharedObjects to initialize their state.

In the future, checkpoints of this materialized view can be saved in
something like ValKey, which will prevent consumers from having to read
through the entire topic to understand the current state of each object.

This commit also adjusts some tests to reduce their flakyness.

Signed-off-by: Ryan Drew <learnitall0@gmail.com>